Add wallet screens first draft
Just a simple first version of add wallet screens. Plenty of stuff missing there still. But it's a prototype, so it's OK.

Also putting nav expand and theme in local storage, so it's easier to dev (doesn't reset on refreshes).

And starting to move some pages into sub-folders, as the number of pages is getting quite huge.